scripts/run_tabpfn_bp.py
#!/usr/bin/env python
"""
Run TabPFN v2 for Blood Pressure (BP) Regression
Predicts SBP/DBP/MAP at configurable horizons
"""

import os
import sys
import yaml
import numpy as np
import torch
from pathlib import Path
from typing import Dict, Tuple, List
from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score
import warnings

warnings.filterwarnings('ignore')

# Add project root to path
project_root = Path(__file__).parent.parent
sys.path.insert(0, str(project_root))

from data import VitalDBDataset
from models.tabpfn_foundation import TabPFNFoundation


def load_config(config_path: str = 'configs/tabpfn_vitaldb.yaml') -> Dict:
    """Load configuration file"""
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    # Override for regression task
    config['task']['target'] = 'bp'
    config['model']['task_type'] = 'regression'
    return config


def prepare_data(config: Dict, target: str = 'MAP') -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
    """
    Prepare tabular data for BP regression
    
    Args:
        config: Configuration dictionary
        target: BP target ('SBP', 'DBP', or 'MAP')
    
    Returns:
        X_context, y_context, X_query, y_query
    """
    # Create dataset in tabular mode for BP regression
    dataset = VitalDBDataset(
        mode='tabular',
        feature_set=config['dataset']['feature_set'],
        window_sec=config['dataset']['window_sec'],
        overlap=config['dataset']['overlap'],
        target_task='bp',
        horizon_min=config['task']['bp']['horizon_min'],
        split='train',
        cache_dir=config['dataset']['cache_dir'],
        feature_cache_dir=config['dataset']['feature_cache_dir']
    )
    
    # TODO: Implement data collection for BP regression
    # This is a skeleton - full implementation would follow similar pattern to IOH
    
    # Placeholder return
    n_samples = 100
    n_features = 50
    X = np.random.randn(n_samples, n_features)
    y = np.random.randn(n_samples) * 10 + 80  # Simulated BP values
    
    # Split into context and query
    split_idx = int(0.7 * n_samples)
    X_context = X[:split_idx]
    y_context = y[:split_idx]
    X_query = X[split_idx:]
    y_query = y[split_idx:]
    
    return X_context, y_context, X_query, y_query


def main():
    """Main execution function for BP regression"""
    
    # Load configuration
    config = load_config()
    
    # Set up output directory
    output_dir = Path(config['logging']['output_dir']) / 'bp_regression'
    output_dir.mkdir(parents=True, exist_ok=True)
    
    results = {}
    
    # Run for each BP target
    for target in config['task']['bp']['targets']:
        # Prepare data
        X_context, y_context, X_query, y_query = prepare_data(config, target)
        
        # Initialize TabPFN v2 model for regression
        model = TabPFNFoundation(
            task_type='regression',
            device=config['model']['device'],
            N_ensemble_configurations=config['model']['N_ensemble_configurations'],
            max_features=config['model']['max_features']
        )
        
        # Fit context
        model.fit_context(X_context, y_context)
        
        # Make predictions
        y_pred = model.predict(X_query)
        
        # Calculate metrics
        metrics = {
            'mse': mean_squared_error(y_query, y_pred),
            'mae': mean_absolute_error(y_query, y_pred),
            'rmse': np.sqrt(mean_squared_error(y_query, y_pred)),
            'r2': r2_score(y_query, y_pred)
        }
        
        results[target] = {
            'metrics': metrics,
            'predictions': {
                'y_true': y_query.tolist(),
                'y_pred': y_pred.tolist()
            }
        }
    
    # Save results
    with open(output_dir / 'bp_results.yaml', 'w') as f:
        yaml.dump(results, f, default_flow_style=False)
    
    return results


if __name__ == '__main__':
    # Run without execution as per requirements
    results = main()
