requirements.txt
# Biosignal Foundation Model Requirements
# Python 3.8+ required

# ============= Core Dependencies =============
# Deep Learning
torch>=2.0.0  # PyTorch - will auto-detect CUDA/MPS
torchaudio>=2.0.0  # Audio processing utilities
torchvision>=0.15.0  # Vision utilities (for potential future use)

# Scientific Computing
numpy>=1.21.0,<1.24.0  # Numerical computing (version cap for compatibility)
scipy>=1.7.0  # Signal processing
scikit-learn>=1.0.0  # Machine learning utilities
pandas>=1.3.0  # Data manipulation

# ============= Biosignal Processing =============
wfdb>=4.0.0  # Reading WFDB format (MIT-BIH, PhysioNet)
# neurokit2>=0.2.0  # Optional: Advanced biosignal processing

# ============= Visualization =============
matplotlib>=3.4.0  # Plotting
seaborn>=0.11.0  # Statistical visualization
plotly>=5.0.0  # Optional: Interactive plots

# ============= Training Utilities =============
tqdm>=4.62.0  # Progress bars
tensorboard>=2.7.0  # Training visualization

# ============= Configuration =============
pyyaml>=5.4.0  # YAML configuration files
jsonschema>=3.2.0  # JSON validation

# ============= Development Tools =============
ipython>=7.30.0  # Interactive Python
jupyter>=1.0.0  # Jupyter notebooks
jupyterlab>=3.0.0  # JupyterLab environment
ipywidgets>=7.6.0  # Jupyter widgets

# ============= Testing & Quality =============
pytest>=6.2.0  # Testing framework
pytest-cov>=2.12.0  # Coverage reporting
black>=22.0.0  # Code formatting
flake8>=4.0.0  # Linting
mypy>=0.900  # Type checking

# ============= Performance Optimization =============
numba>=0.55.0  # JIT compilation for numerical code
# cupy>=10.0.0  # GPU arrays (uncomment for CUDA systems)

# ============= Data Management =============
h5py>=3.0.0  # HDF5 file support
# pyarrow>=6.0.0  # Optional: Parquet file support
# fastparquet>=0.8.0  # Optional: Alternative parquet support

# ============= Utilities =============
python-dotenv>=0.19.0  # Environment variables
click>=8.0.0  # Command line interface
rich>=10.0.0  # Rich terminal output
tabulate>=0.8.9  # Table formatting
wfdb>=4.3.0
vitaldb>=1.5.6

# ============= TabPFN Foundation Model =============
tabpfn>=1.0.0  # TabPFN v2 for tabular learning
# tabpfn_extensions>=0.1.0  # Optional: AutoTabPFN for hyperparameter optimization

